Edit & Delete User Restrictions

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,14 +92,31 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model ,HttpSession session) {
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
-    }
+        User loggedUser = (User)session.getAttribute("loggeduser");
+
+        String tags = "" ;
+
+
+
+        if(image.getUser().getId() == loggedUser.getId()){
+            model.addAttribute("image", image);
+            if(!image.getTags().isEmpty()) {
+                tags = convertTagsToString(image.getTags());
+            }
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }else{
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError",true);
+            return "images/image";
+        }
+
+
+}
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
     //The method receives the imageFile, imageId, updated image, along with the Http Session
@@ -132,7 +149,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/"+updatedImage.getId()+"/"+ updatedImage.getTitle();
     }
 
 
@@ -140,9 +157,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        User loggedUser = (User)session.getAttribute("loggeduser");
+
+        Image image = imageService.getImage(imageId);
+        if(image.getUser().getId() == loggedUser.getId()) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else{
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError",true);
+            return "images/image";
+        }
+
     }
